move external indexing to separate file, add read/write timeouts for socket, close socket on interrupt
